+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased 2.1]
+## Version 2.1.0
+
+- XML-based routing configuration has been deprecated in Symfony 7.4. Switch to PHP-based routing configuration by including
+  `guide-routing.php` instead of `guide-routing.xml`. A deprecation notice will be triggered.
 
 ## Version 2.0.0